+++ R/formatting_properties.R
@@ -361,7 +361,7 @@ print.fp_border <- function(x, ...) {
 #' a valid color (e.g. "#000000" or "black").
 #' @param keep_with_next a scalar logical. Specifies that the paragraph (or at least part of it) should be rendered
 #' on the same page as the next paragraph when possible.
-#' @param word_style_id Word paragraph style identifier
+#' @param word_style Word paragraph style name
 #' @return a \code{fp_par} object
 #' @examples
 #' fp_par(text.align = "center", padding = 5)
@@ -378,7 +378,7 @@ fp_par = function(text.align = "left",
                   border.top, border.right,
                   shading.color = "transparent",
                   keep_with_next = FALSE,
-                  word_style_id = "Normal") {
+                  word_style = "Normal") {
 
   out = list()
 
@@ -414,7 +414,7 @@ fp_par = function(text.align = "left",
   if( !missing(border.right) )
     out <- check_set_border( obj = out, border.right)
 
-  out <- check_set_chr(obj = out, word_style_id)
+  out <- check_set_chr(obj = out, word_style)
 
   out$keep_with_next <- keep_with_next
   class( out ) = "fp_par"
@@ -477,13 +477,13 @@ print.fp_par = function (x, ...){
 update.fp_par <- function(object, text.align, padding, border,
                           padding.bottom, padding.top, padding.left, padding.right,
                           border.bottom, border.left,border.top, border.right,
-                          shading.color, keep_with_next, word_style_id, ...) {
+                          shading.color, keep_with_next, word_style, ...) {
 
   if( !missing( text.align ) )
     object <- check_set_choice( obj = object, value = text.align,
                                 choices = c("left", "right", "center", "justify") )
-  if( !missing( word_style_id ) ){
-    object <- check_set_chr(obj = object, word_style_id)
+  if( !missing( word_style ) ){
+    object <- check_set_chr(obj = object, word_style)
   }
 
   # padding checking